+++ rust/swift_bindings/ChromaDemo/ChromaDemo/ChromaDemoApp.swift
@@ -1,14 +1,15 @@
 //
-//  ChromaLocalTestApp.swift
-//  ChromaLocalTest
+//  ChromaDemoApp.swift
+//  ChromaDemo
 //
-//  Created by Nicholas Arner on 5/13/25.
+//  Created by Nicholas Arner on 5/14/25.
 //
 
 import SwiftUI
+import Chroma
 
 @main
-struct ChromaLocalTestApp: App {
+struct ChromaDemoApp: App {
     var body: some Scene {
         WindowGroup {
             ContentView()
