+++ rust/swift_bindings/ChromaDemo/ChromaDemoTests/ChromaDemoTests.swift
@@ -1,14 +1,13 @@
 //
-//  ChromaLocalTestTests.swift
-//  ChromaLocalTestTests
+//  ChromaDemoTests.swift
+//  ChromaDemoTests
 //
-//  Created by Nicholas Arner on 5/13/25.
+//  Created by Nicholas Arner on 5/14/25.
 //
 
 import Testing
-@testable import ChromaLocalTest
 
-struct ChromaLocalTestTests {
+struct ChromaDemoTests {
 
     @Test func example() async throws {
         // Write your test here and use APIs like `#expect(...)` to check expected conditions.
